refactor(types): extract the EmailAddress type to its own module
