+++ src/lib.rs
@@ -73,6 +73,7 @@ pub mod app_errors;
 pub mod auth_db;
 pub mod bounces;
 pub mod duration;
+pub mod email_address;
 pub mod logging;
 pub mod message_data;
 pub mod providers;
